requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
category_encoders==2.7.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
choreographer==1.1.1
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
Cython==3.1.4
dash==3.2.0
decorator==5.2.1
deprecation==2.1.0
dotenv==0.9.9
executing==2.2.1
fastapi==0.115.12
fastjsonschema==2.21.2
Flask==3.1.2
fonttools==4.60.0
gitdb==4.0.12
GitPython==3.1.45
google-genai==1.17.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
imbalanced-learn==0.14.0
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kaleido==1.1.0
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.45.0
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
narwhals==2.5.0
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.62.0
numpy==1.26.4
oracledb==3.1.0
orjson==3.11.3
packaging==24.2
pandas==2.1.4
parso==0.8.5
patsy==1.0.1
pillow==11.3.0
platformdirs==4.4.0
plotly==6.1.2
plotly-resampler==0.11.0
pluggy==1.6.0
pmdarima==2.0.4
prompt_toolkit==3.0.52
protobuf==5.29.5
psutil==7.1.0
pure_eval==0.2.3
pyarrow==21.0.0
pycaret==3.3.2
pycparser==2.23
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyod==2.0.5
pyparsing==3.2.5
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
retrying==1.4.2
rpds-py==0.27.1
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
simplejson==3.20.1
six==1.17.0
sktime==0.26.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.5
streamlit==1.44.1
tbats==1.1.3
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xxhash==3.5.0
yellowbrick==1.5
zipp==3.23.0
psycopg2-binary==2.9.10
boto3==1.40.41

# Pacotes de teste
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-asyncio==0.24.0
memory-profiler==0.61.0
httpx==0.28.1
objgraph==3.6.2
